@@ -2,55 +2,45 @@
 
 import pathlib
 
-from tilda import VERSION
 from setuptools import find_packages, setup
 
 PROJECT_ROOT = pathlib.Path(__file__).parent
-README = (PROJECT_ROOT / 'README.md').read_text()
-
-ldap = ['freenit[ldap]']
-mongo = ['freenit[mongoengine]']
-sql = ['freenit[sql]']
-dev = ['freenit[dev]']
+README = (PROJECT_ROOT / "README.md").read_text()
 
 setup(
-    name='tilda',
-    version=VERSION,
-    description='REST API framework based on Flask-Smorest',
+    name="tilda",
+    version="0.0.1",
+    description="REST API based on Freenit",
     long_description=README,
-    long_description_content_type='text/markdown',
-    url='https://github.com/freenit-framework/backend',
-    author='Goran Mekić',
-    author_email='meka@tilda.center',
-    license='BSD',
+    long_description_content_type="text/markdown",
+    url="https://github.com/organization/backend",
+    author="John Doe",
+    author_email="john@example.com",
+    license="BSD",
     classifiers=[
-        'Development Status :: 1 - Planning',
-        'Intended Audience :: Developers',
-        'Topic :: Internet :: WWW/HTTP',
-        'Environment :: Web Environment',
-        'Framework :: Flask',
-        'License :: OSI Approved :: BSD License',
-        'Programming Language :: Python :: 3',
+        "Development Status :: 4 - Beta",
+        "Intended Audience :: Developers",
+        "Topic :: Internet :: WWW/HTTP",
+        "Environment :: Web Environment",
+        "Framework :: FastAPI",
+        "License :: OSI Approved :: BSD License",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
     ],
     keywords=[
-        'REST',
-        'openapi',
-        'swagger',
-        'flask',
-        'marshmallow',
-        'apispec'
-        'webargs',
+        "REST",
+        "openapi",
+        "swagger",
+        "fastapi",
     ],
-    packages=find_packages(exclude=['tests*']),
-    python_requires='>=3.5',
-    install_requires=[
-        'freenit[ldap]',
-    ],
+    packages=find_packages(exclude=["tests*"]),
+    python_requires=">=3.8",
+    install_requires=["freenit[ormar]"],
     extras_require={
-        'all': ldap + mongo + sql,
-        'ldap': ldap,
-        'mongo': mongo,
-        'sql': sql,
-        'dev': dev,
+        "build": ["freenit[build]"],
+        "dev": ["freenit[dev]"],
+        "test": ["freenit[test]"],
     },
 )
